# build: align debugging flags to `-O0`