--- a/depends/hosts/mingw32.mk
+++ b/depends/hosts/mingw32.mk
@@ -17,7 +17,7 @@ endif
 mingw32_release_CFLAGS=-O2
 mingw32_release_CXXFLAGS=$(mingw32_release_CFLAGS)
 
-mingw32_debug_CFLAGS=-O1 -g
+mingw32_debug_CFLAGS=-O0 -g
 mingw32_debug_CXXFLAGS=$(mingw32_debug_CFLAGS)
 
 mingw32_debug_CPPFLAGS=-D_GLIBCXX_DEBUG -D_GLIBCXX_DEBUG_PEDANTIC