--- a/depends/hosts/netbsd.mk
+++ b/depends/hosts/netbsd.mk
@@ -10,7 +10,7 @@ endif
 netbsd_release_CFLAGS=-O2
 netbsd_release_CXXFLAGS=$(netbsd_release_CFLAGS)
 
-netbsd_debug_CFLAGS=-O1 -g
+netbsd_debug_CFLAGS=-O0 -g
 netbsd_debug_CXXFLAGS=$(netbsd_debug_CFLAGS)
 
 ifeq (86,$(findstring 86,$(build_arch)))