--- a/depends/hosts/openbsd.mk
+++ b/depends/hosts/openbsd.mk
@@ -4,7 +4,7 @@ openbsd_CXXFLAGS=-pipe -std=$(CXX_STANDARD)
 openbsd_release_CFLAGS=-O2
 openbsd_release_CXXFLAGS=$(openbsd_release_CFLAGS)
 
-openbsd_debug_CFLAGS=-O1 -g
+openbsd_debug_CFLAGS=-O0 -g
 openbsd_debug_CXXFLAGS=$(openbsd_debug_CFLAGS)
 
 ifeq (86,$(findstring 86,$(build_arch)))