A depends build with `DEBUG=1` and using `cmake -B build -DCMAKE_BUILD_TYPE=Debug` do not align, which is inconsistent/confusing. The depends build will use `-O1`, while `CMAKE_BUILD_TYPE=Debug` will set `-O0`.

I'm not completely sure yet what the right choice is (previously it seems that `-Og` was unusable with Clang? (note that for Clang it looks like `-Og` and `-O1` are basically equivalent), but we should at least align the two systems to be using the same thing for debugging.

Configure history:
`-O0` added in https://github.com/bitcoin/bitcoin/pull/3833
Switched to `-Og` in https://github.com/bitcoin/bitcoin/pull/13005.
Switched back to `-O0` https://github.com/bitcoin/bitcoin/pull/16435

Depends `DEBUG=1` mode has used `-O1` since inception.