@hebasto 2024-04-03
> * A depends build with `DEBUG=1` and using `./configure --enable-debug` do not align, which is inconsistent/confusing.

Concept ACK on fixing that.