@theuni 2024-04-03
Concept ACK.

I think msan is a good proxy for what we want enabled. [From its docs](https://releases.llvm.org/18.1.1/tools/clang/docs/MemorySanitizer.html):

> To get a reasonable performance add -O1 or higher. To get meaningful stack traces in error messages add -fno-omit-frame-pointer. To get perfect stack traces you may need to disable inlining (just use -O1) and tail call elimination (-fno-optimize-sibling-calls).


From [gcc's docs](https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html) on `-Og`

>  -Og should be the optimization level of choice for the standard edit-compile-debug cycle, offering a reasonable level of optimization while maintaining fast compilation and a good debugging experience. It is a better choice than -O0 for producing debuggable code because some compiler passes that collect debug information are disabled at -O0. Like -O0, -Og completely disables a number of optimization passes so that individual options controlling them have no effect. Otherwise -Og enables all -O1 optimization flags except for those that may interfere with debugging

From what I can tell, `-O1` (and thus `-Og`) will disable the frame pointer. So we may also want to add `-fno-omit-frame-pointer` as recommended by llvm, but IMO we can wait and see if it turns out to be necessary for anyone first.